Order and OrderItem: Added more test, changes to services and controllers

+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/GroceryStoreBackend/service/OrderItemService.java
@@ -6,6 +6,7 @@ import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ca.mcgill.ecse321.GroceryStoreBackend.dao.OrderItemRepository;
+import ca.mcgill.ecse321.GroceryStoreBackend.dao.OrderRepository;
 import ca.mcgill.ecse321.GroceryStoreBackend.dao.ShoppableItemRepository;
 import ca.mcgill.ecse321.GroceryStoreBackend.model.Order;
 import ca.mcgill.ecse321.GroceryStoreBackend.model.OrderItem;
@@ -18,20 +19,28 @@ public class OrderItemService {
   OrderItemRepository orderItemRepo;
   @Autowired
   ShoppableItemRepository itemRepo;
+  @Autowired
+  OrderRepository orderRepo;
+  
   
   @Transactional
-  public OrderItem createOrderItem(int quantity, String itemName) throws IllegalArgumentException {
+  public OrderItem createOrderItem(Long orderItemId, int quantity, String itemName, Long orderId) throws IllegalArgumentException {
       
+    Order order = orderRepo.findOrderById(orderId);
+    if(order == null) throw new IllegalArgumentException ("Order Item cannot exist without an order. ");
+    
     if(quantity < 0) throw new IllegalArgumentException ("Please enter a valid quantity. ");
     
     ShoppableItem item = itemRepo.findByName(itemName);
     
     if(item== null) throw new IllegalArgumentException ("Please enter a valid item. ");
-   
+
+    if(orderItemRepo.existsById(orderItemId) == true) throw new IllegalArgumentException ("This ID is being used. ");
     
       OrderItem orderItem = new OrderItem();
       orderItem.setQuantity(quantity);
-      orderItem.setItem(item); 
+      orderItem.setItem(item);
+      orderItem.setId(orderItemId);
       
       orderItemRepo.save(orderItem);
       return orderItem;
@@ -39,7 +48,10 @@ public class OrderItemService {
   }
   
   @Transactional
-  public OrderItem updateOrderItem(Long orderItemId, int quantity, String itemName) throws IllegalArgumentException {
+  public OrderItem updateOrderItem(Long orderItemId, int quantity, String itemName, Long orderId) throws IllegalArgumentException {
+    
+    Order order = orderRepo.findOrderById(orderId);
+    if(order == null) throw new IllegalArgumentException ("Order Item cannot exist without an order. ");
     
     if(quantity < 0) throw new IllegalArgumentException ("Please enter a valid quantity. ");
     
@@ -86,15 +98,20 @@ public class OrderItemService {
   
   
   @Transactional
-  public OrderItem getOrderItemById(Long id) {
+  public OrderItem getOrderItemById(Long orderItemId, Long orderId) {
     
-    if(id== null) throw new IllegalArgumentException ("Please enter a valid order ID. ");
-    OrderItem orderItem = orderItemRepo.findOrderItemById(id);
+    Order order = orderRepo.findOrderById(orderId);
+    if(order == null) throw new IllegalArgumentException ("Order Item does not exist without an order. ");
+
+    if(orderItemId == null) throw new IllegalArgumentException ("Please enter a valid order ID. ");
+    OrderItem orderItem = orderItemRepo.findOrderItemById(orderItemId);
     
     if(orderItem== null) throw new IllegalArgumentException ("Please enter a valid order item by providing a valid order item ID. ");
 
     
-    return orderItem;  }
+    return orderItem;
+
+  }
